+++ bdc_collectors/scripts/load_scihub_provider.sql
@@ -1,5 +1,0 @@
-INSERT INTO bdc.providers (name, description, uri, credentials)
-     VALUES ('SciHub', '',
-             'https://scihub.copernicus.eu/dhus',
-             '{"username": "user@email.com", "password": "password"}')
-         ON CONFLICT DO NOTHING;